requirements.txt
pymongo
python-dotenv
dnspython
SQLAlchemy
Werkzeug
gunicorn
Jinja2
flask
Flask
certifi